track.sh: sanity check file types

+++ track.sh
@@ -12,6 +12,12 @@ echo "Source: $source"
 echo "Destination: $dest"
 echo "Script directory: $script_dir"
 
+echo "Sanity checking file type..."
+if [[ ! -f "$source" ]]; then
+    echo "$source is not a regular file! Terminating." >&2
+    exit 1
+fi
+
 mv "$source" "$dest"
 ln -s "$dest" "$source"
